requirements.txt
appnope==0.1.3
APScheduler==3.10.4
asttokens==2.4.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
DateTime==5.5
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
idna==3.6
install==1.3.5
ipykernel==6.27.1
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil @ file:///private/tmp/python-psutil-20231015-5378-1ng7p6p/psutil-5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
zope.interface==6.2
